+++ lib/hal/HalDisplay.cpp
@@ -37,6 +37,8 @@ void HalDisplay::begin(bool seamless) {
   }
 }
 
+void HalDisplay::setX3SourceDriveLevel(uint8_t vdhVdl) { einkDisplay.setX3SourceDriveLevel(vdhVdl); }
+
 void HalDisplay::clearScreen(uint8_t color) const { einkDisplay.clearScreen(color); }
 
 void HalDisplay::drawImage(const uint8_t* imageData, uint16_t x, uint16_t y, uint16_t w, uint16_t h,